Fix hunyuan_video i2v loader: cast timestep to bfloat16 instead of long

+++ hunyuan_video/pytorch/loader.py
@@ -170,7 +170,7 @@ class ModelLoader(ForgeModel):
             batch_size, image_seq_len, config.image_embed_dim, dtype=dtype
         )
 
-        timestep = torch.tensor([500], dtype=torch.long).expand(batch_size)
+        timestep = torch.tensor([500], dtype=dtype).expand(batch_size)
 
         return {
             "hidden_states": hidden_states,
